Add files via upload
Preprocessor cardinal direction's as 0 fix

Previous iterations had an uninteded side effect of how radius was calculated
and converted to a ring/LED number away from center. To explain it visually,
a circle was drawn around the outside of an image and then based on where
a pixel was along that line, it was decided to be part of a ring belonging to an
LED. This meant the cardinal directions had blank space at the edge. The
circle is now drawn from center to side instead of center to corner.

Important note: This means we can use square images