Add files via upload
SD_read order 
SD_read has been almost completely rewritten. Previously it would
read from the sd card one character at a time until it hit a white space,
then it would atoi the result. The preprocessor formatting has been changed
so that atoi is no longer needed, it reads in a uint8_t directly. In addition, as
many values as there are 3*LEDs are being read in at a time, giving the entire
time slice. There is also no longer a max brightness calculation, it is done in
the preprocessor.