Add files via upload
Preprocessor does MAX_BRIGHTNESS calculation and no longer outputs ascii characters

A MAX_BRIGHTNESS calculation now happens in this file before it is output to a .txt. This saves
precious cycle time on the arduino nano, as reducing the brightness is typically an expensive
division operation. MAX_BRIGHTNESS is a constant set near NUM_LED and NUM_SLICE.
In addition, the format specifier for the output has changed. it is now "%c%c%c" There are no
spaces and it outputs whatever the value is a character instead of a readable format. This is
easier to read from the SD card for the arduino since it doesn't have to do any format checking,
one byte is one byte is one value.